Resolve a circular dependency of templates.
BulletMover now stores a pointer on its body.
