--- src/IEntity.cpp
@@ -1,11 +1,0 @@
-#include "IEntity.h"
-#include "systems/EntitySystem.h"
-
-namespace Sigma {
-	EntitySystem* IEntity::entitySystem = 0;
-
-	template<class T>
-	T* IEntity::GetAs() {
-		return entitySystem->getComponent<T>(this);
-	}
-}
