+++ src/entities/BulletMover.cpp
@@ -2,11 +2,10 @@
 
 #include <bullet/btBulletDynamicsCommon.h>
 #include "components/PhysicalWorldLocation.h"
-#include "systems/EntitySystem.h"
 
 namespace Sigma {
 
-	BulletMover::BulletMover(const id_t entityID, const std::vector<Property>& properties) : IEntity(entityID) {
+	BulletMover::BulletMover(const id_t entityID, const std::vector<Property>& properties) : IEntity(entityID), body(0) {
         IEntity::entitySystem->addFeature(this, FeatureID("PhysicalWorldLocation"), properties);
         IEntity::entitySystem->addFeature(this, FeatureID("InterpolatedMovement"), properties);
         IEntity::entitySystem->addFeature(this, FeatureID("ControllableMove"), properties);
@@ -17,6 +16,8 @@ namespace Sigma {
 	void BulletMover::InitializeRigidBody(const std::vector<Property>& properties) {
 		// Add the body component
 		IEntity::entitySystem->addFeature(this, FeatureID("RigidBody"), properties);
+		// store the body component
+		body = GetAs<Body>()->get();
 		// TODO: move this. Is it really necessary anyway ?
         RigidBody::getBody(entityID).lock()->setActivationState(DISABLE_DEACTIVATION);
 	}
